refactor(Block) : move Block component to ScrollView folder
